add long long int reader

+++ Fall10/NtuplePackage/src/treeReader.cc
@@ -167,7 +167,7 @@ std::vector<int>* treeReader::GetInt(const std::string &name)
   else return new std::vector<int>;
 }
 
-std::vector<int>* treeReader::GetLongLongInt(const std::string &name)
+std::vector<long long int>* treeReader::GetLongLongInt(const std::string &name)
 {
   std::map<std::string,std::vector<long long int> * >::const_iterator it_LLI = m_LLIvectors.find(name);
   if( it_LLI != m_LLIvectors.end() ) return m_LLIvectors[name];
